Rename: 컴포넌트명 및 경로 변경
